requirements.txt
Pillow==8.4.0
SpeechRecognition==3.8.1
pyttsx3==2.90
requests==2.26.0
cohere==2.0.0
sumy==0.8.1
python-dotenv==0.19.2
requests